requirements.txt
flask>=2.3.0
flask-sqlalchemy>=3.0.0
flask-socketio>=5.3.0
torch>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
requests>=2.31.0
pyserial>=3.5

